Carry last observation forward for International totals
In general we want to leave cumulative totals empty if they haven't been reported yet (e.g. for Spain with its 1-day lag). For location == "International" however, the last total observation dates back to 2020-03-10, which creates a discrepancy between the world total and the sum of all locations after that date. So for International totals, we now carry the last observation (e.g. 696 for total cases) forward until the present day.