Swap series for Italy
A conversation on the GitHub repository (https://github.com/pcm-dpc/COVID-19/issues/864) of the Department of Civil Protection leads us to strongly suspect that people undergoing multiple rounds of testing over the course of the pandemic are not repeatedly counted in these figures. For this reason, since 23 October 2020 we use the "tests performed" series as our series of reference for Italy's testing data.