Use combined OxCGRT dataset
BSG has now created a CSV that reports a single 'combined' value for each indicator where a local policy is treated as a half-step. For instance, school and university closures are recorded as C1=3 if they apply nationwide. If these closures only occured in one state or territory, not across the whole country, we would report C1_combined=2.5.