Set new_cases to NA for Turkey on 2020-12-10
A large spike in cumulative cases was introduced by a data correction in Turkey. After this fix, the cumulative total remains as shown in the data, but the new_cases variable(and the ones derived from it) will be set to NA for Turkey on December 10, 2020, to avoid showing a spike in new cases for Turkey/Asia/World that distorts the whole time series.