Add Northern Cyprus to population file
Northern Cyprus data (cases, deaths, tests, vaccinations) is not included in the data for Cyprus. To calculate accurate per-capita metrics, we therefore changed the population of Cyprus on 28 Apr 2020 in our file to exclude the Northern part. Now, in order to count vaccinations for Northern Cyprus, we're adding a specific population count for it, based on the national statistical yearbook of 2019 (http://www.stat.gov.ct.tr/IST-YIL-2019.aspx).