Reinstate incremental US vax data
The new CDC time series is reported by date of vaccination, and due to long reporting delays it looks like vaccinations have strongly decreased in the last few days (identically to the issue with Sweden's deaths: https://ourworldindata.org/covid-sweden-death-reporting). So we're reverting back to an incremental collection for US vaccinations, which gives us a time series by date of report.