Make a pipeline for Denmark batch vaccination
Interestingly, it seems that the data were modified.

The pipeline works because the table is indeed the one from the PDF,
it’s just that the data from the PDFs are different between the last
processing and today (even for past dates).

I am not sure about the usual process, so I put the processed data in
the git diff. If needed I can remove it from the commit.

Signed-off-by: ValentinMouret <valentin.mou@gmail.com>