Decrease proxy cache invalidation time to 10min
Since we now have Cloudflare in front of Netlify, the effect of `s-maxage` is more prominent as it applies to all requests. 

We don't want Cloudflare to have to check every request with the origin server (Netlify) whether a resource is still fresh, so having _some_ cache time is desirable, I think. We can further decrease it if it causes problems.