Update covid-testing-latest-data-source-details.csv
There seems to be differences between WHO's reports (using the terminology ""tests conducted"") and Ministry of Health's bulletins (using ""samples tested""), possibly due to pooled testing. Ministry of Health also used the terminology "lượt người xét nghiệm" (meaning the number of times people tested), which I still have not been able to figure out what it referred to. For example, [the 17 March 2020 bulletin](https://ncov.moh.gov.vn/web/guest/-/6847912-112) reported that as of 15 March 2021, the country had carried out the RT-PCR test with 2,482,302 samples, equivalent to 3,248,873 times of people tested, of which 2,559 were identified positive. The statistic referring to the times of people tested is higher than the number of samples tested. This means one sample can be tested multiple times? Maybe when a pooled test result is positive or indeterminate, all the specimens in the pool need to be retested individually.