Greenland/vax: fix dose calculation
Given the pie chart at the bottom of https://corona.nun.gl/emner/statistik/antal_vaccinerede, the bar chart at the top of the page shows the number of people with *only* 1 dose, and not the number of people with at least 1 dose (i.e. the two bars are mutually exclusive). Therefore to get people_vaccinated we need to add up the two numbers.