Update Panama.csv
Complete update of the vaccination against COVID-19 in the Republic of Panama from January 20 to April 19, 2021. The figures were corrected appropriately, it was contrasted in the In official sources, duplicate days were eliminated, the Ministry of Health was migrated, and incorrect figures that appeared in the newspaper La Estrella de Panamá were corrected with the Official information from the Ministry of Health. The second doses of April 18 and 19 contained in the official page https://vacunas.panamasolidario.gob./vacunometro were added.