Update Panama.csv
Vaccination update against COVID-19 in the Republic of Panama corresponding to March 30 to April 21, 2021.I strongly request that it be approved as I submit it to maintain the integrity of the information on Google and our word in data. The second doses come from the official site https://vacunas.panamasolidario.gob.pa.