Update Panama.csv
This is the last update of the vaccination process against COVID-19 in Panama, considering that we are using the information from the vaccine meter, I added the information from the previous days that appeared on the vaccine meter to have the most complete information. As reported by the government, the information in the statement and the report will be the same shortly, thank you very much for your attention, I appreciate it very much.