Update Panama.csv
Vaccination update against COVID-19 in the Republic of Panama corresponding to April 28 and 29 reported by the Ministry of Health of Panama. The information from Oxford / AstraZeneca is corrected because although it is true that they arrived in the country on April 17, they were not applied until April 22 when Vice President José Gabriel Carrizo was the first to participate in the voluntary vaccination program with that specific vaccine.